Progressed to multiclass classification.
Instead of just looking at detecting if a number is 5 or not, now
started looking at multiclass classification so we can determine if a
value is one of the ten classes (0 - 9 digits) that the data set
includes.

Used Scikit-Learn to try different classifiers for this scenario, either
classifiers that can natively do multi-class classification (Such as
SGD) or classifiers that are inherently binary (SVM) but are able to
compare against various different classes through one-versus-one or
one-versus-rest strategies.

Also used a RandomForestClassifier to get a more accurate classifier and
showed as much with a ROC AUC.